refactor(fieldmap): move stock label to view context and clean up references
